Updated Prevew structure, move files are necessary
