+++ src/components/section/modals/index.js
@@ -8,6 +8,12 @@ import Entry from './Entry'
 import AddEditListItem from './AddEditListItem'
 import EmptyList from './EmptyList'
 import ListOrEmpty from './ListOrEmpty'
+import BuildField from './BuildField'
+import BuildFieldArea from './BuildFieldArea'
+import BuildSubmitBtn from './BuildSubmitBtn'
+import VerifyResponse from './VerifyResponse'
+import DecodeBodyObject from './DecodeBodyObject'
+import DisplayAlert from './DisplayAlert'
 
 export {
     AddItemTitle,
@@ -19,5 +25,11 @@ export {
     Entry,
     AddEditListItem,
     EmptyList,
-    ListOrEmpty
+    ListOrEmpty,
+    BuildField,
+    BuildFieldArea,
+    BuildSubmitBtn,
+    DecodeBodyObject,
+    DisplayAlert,
+    VerifyResponse
 }
