+++ src/components/section/preivew/index.js
@@ -1,23 +1,11 @@
 import SectionPreview from './SectionPreview'
 import Constant from './Constant'
-import BuildField from './BuildField'
-import BuildItem from './BuildItem'
-import BuildList from './BuildList'
-import BuildSubmitBtn from './BuildSubmitBtn'
-import VerifyResponse from './VerifyResponse'
-import DecodeBodyObject from './DecodeBodyObject'
-import DisplayAlert from './DisplayAlert'
-import BuildFieldArea from './BuildFieldArea'
+import PreviewItem from './PreviewItem'
+import PreviewList from './PreviewList'
 
 export {
-    BuildField,
-    BuildFieldArea,
-    BuildItem,
-    BuildList,
-    BuildSubmitBtn,
-    Constant,
-    DecodeBodyObject,
-    DisplayAlert,
     SectionPreview,
-    VerifyResponse
+    Constant,
+    PreviewItem,
+    PreviewList
 }
